requirements.txt
pytesseract~=0.3.10
textract~=1.6.5
pypdf2~=3.0.1
pandas~=2.0.2
torch~=2.0.1
Pillow~=9.5.0
Flask~=2.3.2
Werkzeug~=2.3.6
transformers~=4.30.2
datasets~=2.12.0
setuptools~=67.8.0
opencv-python>=4.7.0.72
numpy>=1.24.3
accelerate==0.20.3
aiohttp==3.8.4
aiosignal==1.3.1
argcomplete==3.1.1
async-timeout==4.0.2
attrs==23.1.0
beautifulsoup4==4.12.2
boto3==1.26.153
botocore==1.29.153
bpemb==0.3.4
cachetools==5.3.1
certifi==2023.5.7
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.3
cloudpickle==2.2.1
colorama==0.4.6
compressed-rtf==1.0.6
conllu==4.5.2
contourpy==1.1.0
cycler==0.11.0
Deprecated==1.2.14
dill==0.3.6
docx2txt==0.8
ebcdic==1.1.1
et-xmlfile==1.1.0
extract-msg==0.41.5
filelock==3.12.2
flair==0.12.2
fonttools==4.40.0
frozenlist==1.3.3
fsspec==2023.6.0
ftfy==6.1.1
future==0.18.3
gdown==4.7.1
gensim==4.3.1
google-api-core==2.11.0
google-auth==2.20.0
google-cloud-vision==3.4.2
googleapis-common-protos==1.59.1
grpcio-status==1.54.2
grpcio==1.54.2
huggingface-hub==0.15.1
hyperopt==0.2.7
idna==3.4
IMAPClient==2.3.1
inexactsearch==1.0.2
Janome==0.4.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
kiwisolver==1.4.4
langdetect==1.0.9
lxml==4.9.2
MarkupSafe==2.1.3
matplotlib==3.7.1
more-itertools==9.1.0
mpld3==0.5.9
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.14
networkx==3.1
olefile==0.46
openpyxl==3.1.2
packaging==23.1
pdf2image==1.16.3
pdf2txt==0.7.3
pdfminer.six==20221105
pip==23.1.2
poppler-utils==0.1.0
pptree==3.1
praw==7.7.0
prawcore==2.3.0
proto-plus==1.22.2
protobuf==4.23.2
psutil==5.9.5
py4j==0.10.9.7
pyarrow==12.0.1
pyasn1-modules==0.3.0
pyasn1==0.5.0
pycparser==2.21
pycryptodome==3.18.0
pyparsing==3.0.9
PySocks==1.7.1
pyspellchecker==0.7.2
python-dateutil==2.8.2
python-magic-bin==0.4.14
python-magic==0.4.27
python-pptx==0.6.21
pytorch-revgrad==0.2.0
pytz==2023.3
PyYAML==6.0
regex==2023.6.3
requests==2.31.0
responses==0.23.1
rsa==4.9
s3transfer==0.6.1
safetensors==0.3.1
scikit-learn==1.2.2
scipy==1.10.1
segtok==1.5.11
sentencepiece==0.1.99
silpa-common==0.3
six==1.16.0
smart-open==6.3.0
sortedcontainers==2.4.0
soundex==1.1.3
soupsieve==2.4.1
SpeechRecognition==3.10.0
spellchecker==0.4
sqlitedict==2.1.0
sympy==1.12
tabulate==0.9.0
threadpoolctl==3.1.0
tokenizers==0.13.3
torchvision==0.15.2
tqdm==4.65.0
transformer-smaller-training-vocab==0.2.4
typing-extensions==4.6.3
tzdata==2023.3
tzlocal==5.0.1
update-checker==0.18.0
urllib3==2.0.3
wcwidth==0.2.6
websocket-client==1.5.3
wheel==0.40.0
Wikipedia-API==0.5.8
wrapt==1.15.0
xlrd==2.0.1
XlsxWriter==3.1.2
xxhash==3.2.0
yarl==1.9.2
